+++ resources/es/constants.ts
@@ -1,0 +1,35 @@
+export const BLOCKED_WORDS = ['carajo', 'chingar', 'coño', 'joder', 'madre'];
+
+export const BLOCKED_PARTS = [];
+
+export const LETTERS_NOT_ALLOWED_IN_WINNING_WORD = [];
+
+export const DICTIONARIES = {
+    nativeSpeakersFromWikipedia: 600_000_000,
+    spellchecker: {
+        dir: 'Letterpress',
+        shortName: 'Letterpress',
+        fullName: 'Letterpress',
+        url: 'https://github.com/lorenbrichter/Words',
+    },
+    spellcheckerAlt: {
+        dir: 'TypingAid',
+        shortName: 'TypingAid',
+        fullName: 'TypingAid',
+        url: 'https://github.com/ManiacDC/TypingAid/',
+    },
+    winning: {
+        dir: 'FreeDict',
+        shortName: 'FreeDict',
+        fullName: 'FreeDict',
+        url: 'https://freedict.org/',
+    },
+    winningAlt: {
+        dir: 'FrequencyWords',
+        shortName: 'FrequencyWords',
+        fullName: 'FrequencyWords',
+        url: 'https://github.com/hermitdave/FrequencyWords/blob/master/content/2018/es/es_50k.txt',
+    },
+};
+
+export const EASTER_EGG_DAYS = { };