+++ src/components/Panes/Settings/SettingsSources.tsx
@@ -120,7 +120,9 @@ const SettingsSources = () => {
               {': '}
               {t('language.currentLanguage')}
           </h2>
-          <ul className={clsx({ 'list-col-3': (dictionaries.length + 1) % 3 === 0 })}>
+          <ul className={clsx({
+            'list-col-3': (dictionaries.length + 1) % 3 === 0,
+          })}>
               <li>
                   <ButtonTile onClick={() => changePane(Pane.AboutLanguage)}>
                       <IconBookOpen />