+++ src/const.ts
@@ -6,7 +6,7 @@ export const isDev = (window.location.origin || '')?.includes('localhost') || fa
 export const WORD_MAXLENGTH = 15;
 
 // If breaking changes are released you can set a date stamp here and it will block the game
-export const UPDATE_BLOCK_DAILY = '19.01.2024';
+export const UPDATE_BLOCK_DAILY = '12.05.2024';
 
 const SUPPORTED_BUT_NOT_INCLUDED_IN_VIRTUAL_KEY_LINES = [' ', 'delete', 'arrowleft', 'arrowright', 'arrowdown', 'arrowup'];
 
@@ -147,7 +147,6 @@ export const SUPPORTED_DICTIONARY_BY_LANG: {
   de: {
     code: 'de',
     languages: ['de', 'de-DE'],
-    isBeta: true,
     title: 'DIFFLE - das Spiel wie Wordle (auf Deutsch, ohne Zeichenbegrenzung) 🇩🇪',
     keyLinesVariants: KEY_LINES_DE_VARIANTS,
     keyLinesToUse: KEY_LINES_DE_VARIANTS[0].keyLines,
@@ -199,7 +198,6 @@ export const SUPPORTED_DICTIONARY_BY_LANG: {
   fr: {
     code: 'fr',
     languages: ['fr'],
-    isBeta: true,
     title: 'DIFFLE - le jeu similaire à Wordle (en français, sans limite de caractères) 🇫🇷',
     keyLinesVariants: KEY_LINES_FR_VARIANTS,
     keyLinesToUse: KEY_LINES_FR_VARIANTS[0].keyLines,
@@ -226,8 +224,8 @@ export const SUPPORTED_DICTIONARY_BY_LANG: {
     specialCharacters: ['à', 'è', 'é', 'ì', 'í', 'î', 'ò', 'ó', 'ù', 'ú'],
     hasSpecialCharacters: true,
     urls: [
-      { url: 'https://sapere.virgilio.it/parole/vocabolario/{{word}}', name: 'Virgilio Sapere', hasExactMatchAlways: false },
       { url: 'https://www.dizionario-italiano.it/dizionario-italiano.php?parola={{word}}', name: 'Dizionario Italiano', hasExactMatchAlways: false },
+      { url: 'https://sapere.virgilio.it/parole/vocabolario/{{word}}', name: 'Virgilio Sapere', hasExactMatchAlways: false },
     ],
     shareMarker: '🇮🇹 #diffle #diffleit',
   },