+++ CHANGELOG.md
@@ -11,7 +11,14 @@
 
 ---
 
-#### Version 3.21
+#### Version 3.22
+- **New**: TypeScript added in node, shared files between node and app
+- **Fix**: Swear words have been removed from Italian winning words
+- **New**: All langs are using FrequencyWords as part of building winning dictionary
+- **New**: A hint was added informing that yesterday's word could be wrong due to the update
+- **New**: There is a date shown in the header for Easter egg days
+
+#### Version 3.21 (12.05.2024)
 - **New**: Italian game language added
 - **New**: Czech dictionary updated